+++ CHANGELOG.md
@@ -10,6 +10,18 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Added
+
+- Added `NewStreamLogger()` and `NewStreamHandler()` functions that create
+  loggers and handlers that write to an `io.Writer`.
+
+### Changed
+
+- **[BC]** Renamed `NewLogger` to `NewTestLogger`
+- **[BC]** Renamed `NewHandler` to `NewTestHandler`
+
 ## [0.1.1] - 2024-04-09
 
 - Include the elapsed duration since the logger was created in each log message.